harvestDaily/requirements.txt
pandas
boto3
beautifulsoup4
python-dotenv
s3fs
awscli